rake resque:start hangs if anything writes to STDOUT
If any code writes to STDOUT during `rake resque:restart` then capistrano hangs.

I've redirected the output of this command to `/dev/null`, which allows the cap
task to complete.

Yes, this will swallow errors. I'm new to working with `capistrano` in this way
and open to suggestions on how to handle that in a more friendly way.

In the meantime, this branch might help some people have cleaner deploys.