Fire-and-forget option for starting workers
...and the scheduler. Where a large number of workers needs to be
started on a single host, it can be a long wait for each one to start
before moving onto the next. This change offers a resque_fire_and_forget
variable which will background the `nohup rake` command.

This does of course have the disadvantage of ignoring any errors that
might occur when executing rake, so this should only be enabled if the
benefits outweigh that cost. It may also spike the load on the host as
it tries to concurrently start as many processes as it can.